test-finland added to admin build

+++ build_docker_images.sh
@@ -4,6 +4,15 @@ if [ "$1" = "admin" ]
 then
   [ -d palace_circulation_admin_ui ] && rm -rf palace_circulation_admin_ui
   git clone --branch=main ssh://git.lingsoft.fi/home/git/palace_circulation_admin_ui.git
+  cd palace_circulation_admin_ui
+  npm i
+  npm run test-finland
+  retval=$?
+  if [ $retval -ne 0 ]; then
+    echo "Admin UI test failed"
+	exit $retval
+  fi
+  cd ..
   docker compose --progress=plain build admin
   [ -d palace_circulation_admin_ui ] && rm -rf palace_circulation_admin_ui
   docker create --name temp-admin "${PWD##*/}"-admin:latest
@@ -17,6 +26,3 @@ else
   ./build_docker_images.sh admin
   docker compose --progress=plain build webapp scripts pg minio os
 fi
-
-
-
